Arranque
# Arranque

This repository is a work in progress. Its purpose is two-fold:

- Provide a context for gathering and reviewing material related to future projects.
- Act as a testing/training ground for the development of new skills and capacities.

I will be building out a variety of files, content and resources over time. When I 
acquired improved skills and familiarity with the toolkits used in this platform, I 
will attempt to generate indices, lists, and material which may find useful in their 
own pursuit of knowledge and skill-building.